+++ app/models/sidekiq_publisher/job.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require "active_record"
-
 module SidekiqPublisher
   class Job < ActiveRecord::Base
     self.table_name = "sidekiq_publisher_jobs"