+++ lib/sidekiq_publisher/engine.rb
@@ -1,10 +1,8 @@
 # frozen_string_literal: true
 
 module SidekiqPublisher
-  class Railtie < Rails::Railtie
-    rake_tasks do
-      load "sidekiq_publisher/tasks.rake"
-    end
+  class Engine < Rails::Engine
+    isolate_namespace SidekiqPublisher
 
     initializer "sidekiq_publisher.configure" do
       SidekiqPublisher.logger = Rails.logger